Removes hard dependency on Encode + Decode
My personal use-case for this library necessitates Encode and Decode for
identifiers and for values. However, that's not a necessary workflow,
and it doesn't make sense to force that requirement on all users.

As an aside, since Encode + Decode are no longer unconditionally
necessary, the MarkovValue trait has ceased to exist.